fix: change data type

+++ exercises/exercise1.py
@@ -11,10 +11,10 @@ airportModel = {
     "column_4": sql.types.String,
     "column_5": sql.types.String,
     "column_6": sql.types.String,
-    "column_7": sql.types.Double,
-    "column_8": sql.types.Double,
+    "column_7": sql.types.Float,
+    "column_8": sql.types.Float,
     "column_9": sql.types.Integer,
-    "column_10": sql.types.Double,
+    "column_10": sql.types.Float,
     "column_11": sql.types.String,
     "column_12": sql.types.String,
     "geo_punkt": sql.types.String
